Initial Support for "Implied" Extensions
Implied extensions allow users to use a special extension URL scheme to reference an element from another version of FHIR. To support this, SUSHI needs to be able to load "supplemental" versions of FHIR in addition to the primary version.  These supplemental FHIR definitions are needed to properly build representative extension definitions on the fly.

This commit introduces basic support but is missing support for some advanced cases like types that don't exist in the current version of FHIR (e.g., CodeableReference), resources that changed names (e.g., MedicationOrder), and elements that themselves have extensions.

This also includes changes to imports and specifically the HasId and HasName classes to address issues I encountered with circular import graphs.