Import CodeSystem definitions with hierarchical codes
When defining a concept to add to a CodeSystem, it is possible to place
it in another concept's list of contained concepts. To do so, list all
the codes that precede it in the hierarchy. The preceding codes must be
defined before being used in a hierarchy. An error is thrown when any
codes from a hierarchy are missing.