Update TestFisher to handle other FHIR versions
The TestFisher will load resources from the FHIR cache if it doesn't find them in the test package. Now that we have test packages for other versions of FHIR, we need to update TestFisher to pull from the right FHIR cache and store to the right test package.

This also updates the R5 test package w/ the necessary additional definitions.