CodeSystem contains caret rules on concepts
A new type of rule is added in order to set internal values, as a
CaretValueRule does, on concepts within a CodeSystem's hierarchy. The
new CodeCaretValueRule is added to the grammar and as a Rule class. This
rule should only be used on CodeSystems and on RuleSets inserted into
CodeSystems. Rather than traversing through the CodeSystem with a path
as a CaretValueRule would, the traversal is specified as a sequence of
one or more codes.

Note that this change only implements importing CodeSystems containing
CodeCaretValueRules. These rules do not currently have any effect on the
exported CodeSystem.