Import CodeSystems with hierarchical codes (#809)
* Import CodeSystem definitions with hierarchical codes

When defining a concept to add to a CodeSystem, it is possible to place
it in another concept's list of contained concepts. To do so, list all
the codes that precede it in the hierarchy. The preceding codes must be
defined before being used in a hierarchy. An error is thrown when any
codes from a hierarchy are missing.

* CodeSystem contains caret rules on concepts

A new type of rule is added in order to set internal values, as a
CaretValueRule does, on concepts within a CodeSystem's hierarchy. The
new CodeCaretValueRule is added to the grammar and as a Rule class. This
rule should only be used on CodeSystems and on RuleSets inserted into
CodeSystems. Rather than traversing through the CodeSystem with a path
as a CaretValueRule would, the traversal is specified as a sequence of
one or more codes.

Note that this change only implements importing CodeSystems containing
CodeCaretValueRules. These rules do not currently have any effect on the
exported CodeSystem.

* Add test for CodeCaretValueRule constructor

* Specifying system inside a hierarchy is an error