Add toFSH to CodeCaretValueRule
CodeCaretValueRule was added by a recent update, and now it has a toFSH
method just like all rules should.
CaretValueRule will fshify string values before using them in toFSH.
ContainsRule will include a cardinality in toFSH. The included
cardinality of "0.." is safe, and will not actually affect any change.
In actual SUSHI usage, there will always be a CardRule that comes next
that sets the correct cardinality on the newly-created slice. But,
without the included cardinality, the ContainsRule would not be valid
FSH.
Change spelling of "preprocessed" in CLI to be consistent with other
places that word appears.

Some of the tests on CodeCaretValueRule use caret paths that don't
represent actual elements on CodeSystem.concept. But, they are still
valid for demonstrating the correctness of the implementation of the
toFSH function.