Add option to output preprocessed FSH (#804)
* AddElementRule adds toFSH method

This method is used to produce FSH that, when processed, will create an
instance of AddElementRule equivalent to the current instance. Several
functions are added to a new file in the fshtypes directory, as they
will be useful for implementing toFSH methods for other FSH classes.

* AssignmentRule adds toFSH method

* BindingRule adds toFSH method

Test file for BindingRule renamed to match class name.

* Add toFSH to remaining structure rules

Any rule that may appear on a FshStructure has a toFSH method. SUSHI
will not necessarily make use of all of these implementations, but all
of them are implemented for consistency and for potential value to other
application developers.

* Add toFSH to FshStructure types

Profile, Extension, Logical, and Resource have toFSH methods provided.
The implementations are similar, but adding the implementation at the
parent class would necessitate a toFSH implementation on the Rule class.
Maybe the Rule class should be abstract?

* Add toFSH on FshValueSet and ValueSetComponentRule

SUSHI does not directly instantiate the parent ValueSetComponentRule
class, but an implementation is provided for completeness.

* Add toFSH on FshCodeSystem and ConceptRule

* Update Jest to try to fix import problems

Module imports are not working correctly during automated tests. Update
jest and ts-jest in the hopes of resolving those problems.

* Add toFSH on FSH Instance

* Add toFSH on Invariant

* Add toFSH on Mapping

* Handle NaN in CardRule

* Add option for producing preprocessed FSH

The -p option will cause SUSHI to output preprocessed FSH. These files
will match their original directory structure. They are placed in a
directory named "_preprocessed" inside the normal output directory.

Preprocessed FSH has aliases resolved and rule set insertions applied.
It will typically be more verbose than the original FSH, but should be
semantically equivalent. This makes it useful for troubleshooting
problems that occur when other programs process the FHIR that SUSHI has
produced.

* Avoid circular imports from Assignment and CaretValue rules

Importing entire directories instead of individual files lead to cases
where files imports would form a circle. Although this wasn't a problem
for the compiler, it presented a problem for automated testing. By using
more specific import statements, these import cycles are removed.

To support more specific imports, some functions and classes are
reorganized to be in their own file rather than in a general-purpose
file containing commonly used functions and classes.

Add the new HasFlags mixin for use by FlagRule and AddElementRule.
Add support for bigint and FshCanonical values when producing FSH for an
AssignmentRule or CaretValueRule.
Fix bug with CardRule producing incorrect FSH when min == 0.

* Add tests for writePreprocessedFSH

Emit info message only once after all files have been written.

* Add toFSH to CodeCaretValueRule

CodeCaretValueRule was added by a recent update, and now it has a toFSH
method just like all rules should.
CaretValueRule will fshify string values before using them in toFSH.
ContainsRule will include a cardinality in toFSH. The included
cardinality of "0.." is safe, and will not actually affect any change.
In actual SUSHI usage, there will always be a CardRule that comes next
that sets the correct cardinality on the newly-created slice. But,
without the included cardinality, the ContainsRule would not be valid
FSH.
Change spelling of "preprocessed" in CLI to be consistent with other
places that word appears.

Some of the tests on CodeCaretValueRule use caret paths that don't
represent actual elements on CodeSystem.concept. But, they are still
valid for demonstrating the correctness of the implementation of the
toFSH function.

* Show comment in empty files after preprocessing

A FSH file that contains only entities that do not persist through
preprocessing steps will be empty. When writing this file, put a comment
in it rather than just leaving it blank. This should help make it
clearer why there is no actual FSH content in the file.

* Add -p flag to README

* Include Aliases in preprocessed FSH output

Aliases that are used as slice names do not update the rule paths
that refer to those aliases. Therefore, they should still be included as
part of the preprocessed FSH.