Export CodeSystems with hierarchical codes
Concepts within a code system are arranged hierarchically based on the
ConceptRule's specified hierarchy. The path to a hierarchical code must
exist in order to be added, and this is checked when adding the
ConceptRule to the FshCodeSystem.
Additionally, a CodeCaretValueRule may be used to specify values on a
concept anywhere in the hierarchy. The codePath on this rule is
traversed after all the concepts are added so that the correct element
in the CodeSystem can be located. Just like for CaretValueRules, the
value to set is validated before being assigned.