Set code path contexts when parsing CodeSystems
When defining a CodeSystem, concept definitions establish a code path
context. This is analogous to the path context established by other rule
types that include paths. This allows for the definition of hierarchical
code systems by indenting subsequent concept definitions. This also
allows for CodeCaretValueRules to reference the most recently defined
concept by indenting the CodeCaretValueRule definition.

Because code context paths can now be used, a CodeCaretValueRule
definition may not directly name the concept to which the rule applies.
As a result, the caretValueRule parser rule is no longer a choice for a
csRule. The actual CaretValueRule class, however, is still applicable,
and is created when the visited codeCaretValueRule ends up with no code
path. A convenient side effect is that there is no longer a need to
explicitly check for and prohibit a CaretValueRule with a non-empty path
on a CodeSystem.