Allow rules that establish code path context without creating a concept
Just as the pathRule allows for establishing a path context without
adding any new information, allow for a special case of conceptRule to
establish a code path context without adding any new information. In
order to be considered as a context-setting rule, the conceptRule must
not contain display or definition, and must have a hierarchy that
matches the existing code. If these conditions are met, no error is
emitted. Otherwise, the existing error for duplicate concept definition
is emitted.