Implied Extensions (#805)
* Refactor loadExternalDependencies to return Promise<void>

The existing implementation was strange because it returned an array of promises that all resolved to the same FHIRDefinitions instance as the one that was passed in as an argument.  So the actual values were redundant and useless (and in fact were never used).  So what we really need is a single promise that just resolves when everything is loaded.

* Update TS, Jest, and related libs

I ran into a bug (related to the next commit) that seemed to be an issue w/ the environment, so I upgraded typescript, jest, and related libs.  It turns out the issue was w/ ts-jest only, but it makes sense to go ahead and update them all!

* Initial Support for "Implied" Extensions

Implied extensions allow users to use a special extension URL scheme to reference an element from another version of FHIR. To support this, SUSHI needs to be able to load "supplemental" versions of FHIR in addition to the primary version.  These supplemental FHIR definitions are needed to properly build representative extension definitions on the fly.

This commit introduces basic support but is missing support for some advanced cases like types that don't exist in the current version of FHIR (e.g., CodeableReference), resources that changed names (e.g., MedicationOrder), and elements that themselves have extensions.

This also includes changes to imports and specifically the HasId and HasName classes to address issues I encountered with circular import graphs.

* Implied extensions: Convert unknown types into complex extensions

If I type from the other version of FHIR doesn't exist in this version of FHIR, convert it into a complex extension.

This also includes special logic for R5 CodeableReference to move the binding and reference types to the respective sub-extensions.

* Implied extensions: Replace old resources w/ new resources where applicable

Also warn when an implied extension value type is filtered out because it doesn't exist in this version of FHIR.

* Fix error/warn messages for implied extension support

* Change approach to implied complex extensions

Previously, implied complex extensions simply specified their sub-extensions by referring to the full extension urls for each one.  For example, the DSTU2 Patient.animal extension had a sub-extension pointing to the DSTU2 Patient.animal.species extension.

BUT... this is not how the IG Publisher does it.  The IG publisher fully defines the whole complex extension using relative URLs instead of absolute URLs.  So... now we do the same.

* Add test for fishing for implied extensions

* Cleanly Separate R4 and R5 tests

Don't mix the R4 and R5 definitions in the same FHIRDefinitions object because there are unexpected consequences. (This fixes issues after the merge from master).

* Update TestFisher to handle other FHIR versions

The TestFisher will load resources from the FHIR cache if it doesn't find them in the test package. Now that we have test packages for other versions of FHIR, we need to update TestFisher to pull from the right FHIR cache and store to the right test package.

This also updates the R5 test package w/ the necessary additional definitions.

* Move loadSupplementalFHIRPackage to load.ts

* Address PR feedback

* Fix broken test and prettier violations after merge