Check for invalid hierarchy in export
When exporting a code system, check if a concept's hierarchy actually
exists. If it does not, show an error, and do not add the concept to the
code system. Concepts added normally through FSH will always have valid
hierarchies, but it is possible to add rules directly to a FshCodeSystem
without using the addConcept method.

CodeCaretValueRules don't support soft indexing, so there is no need to
resolve soft indices on those rules.