Export CodeSystems with hierarchical codes (#814)
* Export CodeSystems with hierarchical codes

Concepts within a code system are arranged hierarchically based on the
ConceptRule's specified hierarchy. The path to a hierarchical code must
exist in order to be added, and this is checked when adding the
ConceptRule to the FshCodeSystem.
Additionally, a CodeCaretValueRule may be used to specify values on a
concept anywhere in the hierarchy. The codePath on this rule is
traversed after all the concepts are added so that the correct element
in the CodeSystem can be located. Just like for CaretValueRules, the
value to set is validated before being assigned.

* Check for invalid hierarchy in export

When exporting a code system, check if a concept's hierarchy actually
exists. If it does not, show an error, and do not add the concept to the
code system. Concepts added normally through FSH will always have valid
hierarchies, but it is possible to add rules directly to a FshCodeSystem
without using the addConcept method.

CodeCaretValueRules don't support soft indexing, so there is no need to
resolve soft indices on those rules.