Only create CodeCaretValueRule for CodeSystems
When importing FSH, do not create CaretValueRules on a CodeSystem.
Instead, create a CodeCaretValueRule with an empty code path.
CaretValueRules may still end up on a CodeSystem if a RuleSet containing
CaretValueRules is inserted.

Simplify export of CodeSystems to apply CaretValueRules and
CodeCaretValueRules together, since they want to do the same thing, just
with different base path calculations.