Resolve soft indexing for CodeCaretValueRules
When applying CodeCaretValueRules during a CodeSystem export, the caret
paths on those rules may use soft indexing. First, convert the codePath
into a regular path. That way, the soft index resolver can do what it
always does, which is use the rule's path in order to determine soft
index values. Then, use the combined path and caretPath to assign the
value to the correct element in the CodeSystem.