Support indentation in CodeSystem rules (#818)
* Set code path contexts when parsing CodeSystems

When defining a CodeSystem, concept definitions establish a code path
context. This is analogous to the path context established by other rule
types that include paths. This allows for the definition of hierarchical
code systems by indenting subsequent concept definitions. This also
allows for CodeCaretValueRules to reference the most recently defined
concept by indenting the CodeCaretValueRule definition.

Because code context paths can now be used, a CodeCaretValueRule
definition may not directly name the concept to which the rule applies.
As a result, the caretValueRule parser rule is no longer a choice for a
csRule. The actual CaretValueRule class, however, is still applicable,
and is created when the visited codeCaretValueRule ends up with no code
path. A convenient side effect is that there is no longer a need to
explicitly check for and prohibit a CaretValueRule with a non-empty path
on a CodeSystem.

* Allow rules that establish code path context without creating a concept

Just as the pathRule allows for establishing a path context without
adding any new information, allow for a special case of conceptRule to
establish a code path context without adding any new information. In
order to be considered as a context-setting rule, the conceptRule must
not contain display or definition, and must have a hierarchy that
matches the existing code. If these conditions are met, no error is
emitted. Otherwise, the existing error for duplicate concept definition
is emitted.

* Add assertion for CodeCaretValueRule

All the other rules have an assertion method, so this rule should also
have one.

* Set code path contexts when parsing CodeSystems

When defining a CodeSystem, concept definitions establish a code path
context. This is analogous to the path context established by other rule
types that include paths. This allows for the definition of hierarchical
code systems by indenting subsequent concept definitions. This also
allows for CodeCaretValueRules to reference the most recently defined
concept by indenting the CodeCaretValueRule definition.

Because code context paths can now be used, a CodeCaretValueRule
definition may not directly name the concept to which the rule applies.
As a result, the caretValueRule parser rule is no longer a choice for a
csRule. The actual CaretValueRule class, however, is still applicable,
and is created when the visited codeCaretValueRule ends up with no code
path. A convenient side effect is that there is no longer a need to
explicitly check for and prohibit a CaretValueRule with a non-empty path
on a CodeSystem.

* Allow rules that establish code path context without creating a concept

Just as the pathRule allows for establishing a path context without
adding any new information, allow for a special case of conceptRule to
establish a code path context without adding any new information. In
order to be considered as a context-setting rule, the conceptRule must
not contain display or definition, and must have a hierarchy that
matches the existing code. If these conditions are met, no error is
emitted. Otherwise, the existing error for duplicate concept definition
is emitted.

* Add assertion for CodeCaretValueRule

All the other rules have an assertion method, so this rule should also
have one.

* Only create CodeCaretValueRule for CodeSystems

When importing FSH, do not create CaretValueRules on a CodeSystem.
Instead, create a CodeCaretValueRule with an empty code path.
CaretValueRules may still end up on a CodeSystem if a RuleSet containing
CaretValueRules is inserted.

Simplify export of CodeSystems to apply CaretValueRules and
CodeCaretValueRules together, since they want to do the same thing, just
with different base path calculations.

* Resolve soft indexing for CodeCaretValueRules

When applying CodeCaretValueRules during a CodeSystem export, the caret
paths on those rules may use soft indexing. First, convert the codePath
into a regular path. That way, the soft index resolver can do what it
always does, which is use the rule's path in order to determine soft
index values. Then, use the combined path and caretPath to assign the
value to the correct element in the CodeSystem.

* Combine context-checking code in importer

Checking for validity of a path and a code path shares some logic, so
that logic has been moved to a shared function.

Co-authored-by: Chris Moesel <cmoesel@mitre.org>