Merge CodeCaretValueRule with CaretValueRule
This refactors CodeCaretValueRule out, and adds a pathArray and
isCodeCaretRule flag to CaretValueRule, so that CaretValueRule
can be used to accomplish all the same things as CodeCaretValueRule.
This also allows us to clean up the importer to only have one
context processor. Finally, this fixes bug #834, which was
happening because there was no way to distinguish a CodeCaretValueRule
from a CaretValueRule on a RuleSet. Now there is no need to
distinguish the two.