Merge CodeCaretValueRule with CaretValueRule (#839)
* Merge CodeCaretValueRule with CaretValueRule

This refactors CodeCaretValueRule out, and adds a pathArray and
isCodeCaretRule flag to CaretValueRule, so that CaretValueRule
can be used to accomplish all the same things as CodeCaretValueRule.
This also allows us to clean up the importer to only have one
context processor. Finally, this fixes bug #834, which was
happening because there was no way to distinguish a CodeCaretValueRule
from a CaretValueRule on a RuleSet. Now there is no need to
distinguish the two.

* Update src/import/FSHImporter.ts

Co-authored-by: Julia Afeltra <30803904+jafeltra@users.noreply.github.com>

* Add a test of an indented CaretValueRule below code on RuleSet

* Fix bug with '.' paths

There was an issue where a path with just . was not correctly
processed by the new context logic, since . has a different
meaning when used alone then when used as a separator of paths.
This commit should fix that bug.

Co-authored-by: Julia Afeltra <30803904+jafeltra@users.noreply.github.com>