Warn when unfolding non-choice element with multiple profiles
When assigning a value on an Instance's element, if that element is
nested under a non-choice element with multiple profiles, it cannot be
determined which of the profiles should apply. Therefore, ignore all
profiles and use the base type. Warn the user that this is being done.

Previously in this scenario, the unfolding would not occur, which
resulted in an error being logged indicating that the path to the nested
element did not exist.