Warn when unfolding non-choice element with multiple profiles (#861)
* Warn when unfolding non-choice element with multiple profiles

When assigning a value on an Instance's element, if that element is
nested under a non-choice element with multiple profiles, it cannot be
determined which of the profiles should apply. Therefore, ignore all
profiles and use the base type. Warn the user that this is being done.

Previously in this scenario, the unfolding would not occur, which
resulted in an error being logged indicating that the path to the nested
element did not exist.

* Strengthen test case for non-choice multi-profile

The test case now demonstrates that the profiles are being completely
ignored by assigning a value that would not be valid on either profile,
but is valid on the base type.

Change variable name in ElementDefinition.unfold for easier reading.

* fix typo explaining when to use an available profile

Co-authored-by: Julia Afeltra <30803904+jafeltra@users.noreply.github.com>

Co-authored-by: Julia Afeltra <30803904+jafeltra@users.noreply.github.com>