Fix bug with applying path to insert rules with empty paths
When a rule in a RuleSet has an empty path, and that RuleSet
is applied from an insert rule that has a path context, we
were incorrectly joining the empty path with the context,
resulting in paths such as "name.", which are incorrect. This
fixes that and adds a test.