Remove ig-data folder from IG test fixtures
The ig-data folder represents a deprecated part of the SUSHI project
structure. The tests had already been updated such that this folder was
just an extra part of the file path, without providing any meaningful
separation of files. Removing this folder means that file paths for
these tests and fixtures are now shorter without any changes to
functionality.