Allow aliases with fragments (#880)
* Fix issue with aliases that contain a #

When an alias has a # in it, that causes problems because it gets
parsed as a code, but that is meant to be a sequence. To solve
this, I just made it so that alias can also accept a code. This
feels a bit hacky, but I wasn't sure otherwise how to do it in
the grammar alone.

* Remove unnecessary ? operator and add comment

* Add a comment to the grammar