Adjust error logging strategy for RuleSets with ConceptRules (#881)
A RuleSet can contain ConceptRules and ValueSetConceptComponentRules.
Additionally, it is possible for the same FSH to parse using the concept
or vsConceptComponent parser rules. Therefore, it can be ambiguous when
parsing the RuleSet which rule should actually be made. When one of
these rules is applied, it is handled in different ways depending on
whether the target entity is a ValueSet or a CodeSystem.

If a ConceptRule with a system is part of a RuleSet applied to a
ValueSet, turn it into a ValueSetConceptComponent and apply it. If a
ConceptRule with a system is part of a RuleSet applied to a CodeSystem,
apply it, but log an error, because concepts listed on a CodeSystem
should not include their system. These special cases only apply when the
ConceptRule is part of a RuleSet.