Allow Profiles on Logical Models (#890)
* Allow Profiles on Logical Models

A profile may specify a logical model as its parent. The
StructureDefinition resulting from this will have kind "logical" and
derivation "constraint".

* Distinguish between logical models and profiles of logical models

When adding resources to FHIRDefinitions, both logical models and
profiles of logical models will have "logical" as their kind. Use the
derivation to distinguish between these two resource types.

* Do not create instances of profiles of logical models

If an Instance is defined as an instance of a Profile, and that Profile
is of a logical model, log and error and do not create the Instance.