Avoid duplicate entries in Instance's meta.profile due to versioned profiles (#893)
* Avoid profile duplication due to InstanceOf and versioned profiles

If an Instance's InstanceOf keyword specifies a profile, that profile's
URL is added to the exported resource's meta.profile array. If a
versioned profile with the same URL is also in the meta.profile array,
remove the unversioned profile.

* Keep unversioned profile if added by rule

If the InstanceOf profile on an Instance is added to meta.profile by a
rule, then it should not be removed, even if another rule has added the
versioned InstanceOf profile.

Cleaned up various InstanceExporter tests that had mismatches between
the FSH in comments and the rule definitions.

* Add InstanceOf profile before validating Instance

An Instance may have a minimum cardinality on meta.profile, so
(potentially) add the InstanceOf profile before checking for required
elements. This way, the InstanceOf profile can help contribute towards
that minimum.