Elements can be constrained with canonicals that contain versions (#898)
* OnlyRule canonical types may include version

A version can be included as part of a canonical type in an OnlyRule.
This version is included in the target profiles of the type.

* Assign Canonical to an element constrained with a version

If an element with a canonical type is constrained with versioned types,
it is still possible to assign a value to that element as long as it
matches a target profile. The version provided in the constraint is not
checked when this value is assigned.