Try to detect name conflicts adjacent to ParentNotDefinedError
If the name of an entity in the tank (defined by the author) is the same
as the name of an entity in the FHIR definitions, the entity in the tank
will prevent the entity in the FHIR definitions from being found while
fishing. This can result in somewhat confusing scenarios if, for
example, a ValueSet in the tank has the same name as a Profile in the
FHIR definitions, and another Profile tries to use that name as its
parent. Try to detect that scenario, and when detected, let the author
know that there may be conflicting names.