Add ParentNameConflictError for tank resources blocking desired FHIR
When the parent is not found for a StructureDefinition, attempt to fish
up the parent from the FHIR definitions and the FSH tank. If definitions
are found in both locations, this means that the FSH entity with the
parent's name has a type that can't be used as the parent, which was why
nothing was found when fishing for FHIR. Show an error message that
describes this behavior and recommends using the URL of the resource in
the FHIR definitions.