Warn when a concept is included in a ValueSet more than once (#913)
* Warn when a concept is included in a ValueSet more than once

Including the same concept (with the same code, system, and version) in
a ValueSet multiple times has no effect. It does not cause any
validation errors, but it may indicate a mistake when defining the
ValueSet. Warn the user when this is detected.

* Do not add duplicate concepts to ValueSet

When a concept would be included more than once in a ValueSet, only add
it once. If this causes a component of a ValueSet to contain to
concepts, do not add that component to the ValueSet.