Show error when inserting RuleSet with concepts at a path (#921)
If a RuleSet is inserted at a path, that path is the context for the
rules in the RuleSet. However, a ConceptRule should not have a path as
context. So, if a RuleSet that contains a ConceptRule is inserted at a
path, show an error so that the author knows that this should not be
done. The ConceptRule should still be applied.