Check local code systems when composing ValueSets
When a ValueSet contains a rule that adds specific concepts from a
system, check to see if that system is defined locally. If it is, check
to make sure that the listed concepts are part of the system. If any
codes are not part of the system, log one error message listing all the
codes that were not found in the system.