Insert RuleSet at code path
An insert rule on a CodeSystem or RuleSet can start with a code path.
This allows the code path to act as the context for the rules contained
in the inserted RuleSet. When an InsertRule has a code path, it is
applied to ConceptRules and CaretValueRules in the inserted RuleSet. An
InsertRule with a code path on an entity other than a CodeSystem or
RuleSet is syntactically invalid.