Try to validate usage of codes from local code systems (#930)
* Check local code systems when replacing references

When replacing a reference for a FshCode used in an AssignmentRule,
check to make sure that the code being used is actually present in the
system. If it is not present, log an error.

* Check local code systems when composing ValueSets

When a ValueSet contains a rule that adds specific concepts from a
system, check to see if that system is defined locally. If it is, check
to make sure that the listed concepts are part of the system. If any
codes are not part of the system, log one error message listing all the
codes that were not found in the system.

* Check codes in Instances of CodeSystem

When fishing for a CodeSystem, an Instance of CodeSystem may be
returned. Concepts on an Instance are added through AssignmentRules at
specific paths.

When determining the URL of an Instance, check for AssignmentRules on
the url path.

* Check for concepts added by InsertRules

Concepts may be added to CodeSystems or Instances thereof by inserting a
RuleSet that defines additional concepts. Expand the InsertRules on
these before checking that a code being used is present.

* Check local codes used in ValueSetFilterComponentRules

* Refactor local code system check

Reduce duplicate code by moving the local code system check to a
separate function.

* Only check codes for complete CodeSystems

A CodeSystem's content element has a code indicating whether the system
is complete. If a system is complete, then log an error when a code not
in the system is used. If a system is incomplete, do not log an error.