Insert RuleSet at code path (#942)
* Insert RuleSet at code path

An insert rule on a CodeSystem or RuleSet can start with a code path.
This allows the code path to act as the context for the rules contained
in the inserted RuleSet. When an InsertRule has a code path, it is
applied to ConceptRules and CaretValueRules in the inserted RuleSet. An
InsertRule with a code path on an entity other than a CodeSystem or
RuleSet is syntactically invalid.

* InsertRuleContext contains a code path

When checking if a parser context contains a code path, consider the
presence of a RULESET_REFERENCE or PARAM_RULESET_REFERENCE to be
sufficient. Although this will include insert rules that contain a
regular path, those will also get picked up by containsPathContext, so
there are no consequences.