Find instances of profiles when fishing for profiles
An Instance in a FSHTank represents a Profile if it is a definition
instance of StructureDefinition and has a rule that assigns its
derivation as #constraint. An Instance in a Package represents a Profile
if it is a definition instance with resourceType "StructureDefinition"
and derivation "constraint".

When fishing for FHIR in StructureDefinitionExporter, a not-yet-exported
Instance may be found. Export it using InstanceExporter before fishing
again.