Find instances of extensions when fishing for extensions
An Instance in a FSHTank represents an Extension if it is a definition
instance of StructureDefinition, has a rule that assigns its derivation
as #constraint, and has a rule that assigns its type as Extension. An
Instance in a Package represents an Extensions if it is a definition
instance with resourceType "StructureDefinition", derivation
"constraint", and type "Extension".

Instances that represent Profiles must not have type "Extension".