Only modifier extensions should be used on modifierExtension elements
When handling a ContainsRule on an element whose only type is Extension,
check if the extension being added in the ContainsRule is a modifier
extension. A modifier extension is one where the root element has
isModifier set to true. Log an error when a modifier extension is used
on a path that ends in "extension" and when a non-modifier extension is
used on a path that ends in "modifierExtension".