Find Instances of Profile, Extension, CodeSystem, and ValueSet when fishing (#949)
* Find instances of profiles when fishing for profiles

An Instance in a FSHTank represents a Profile if it is a definition
instance of StructureDefinition and has a rule that assigns its
derivation as #constraint. An Instance in a Package represents a Profile
if it is a definition instance with resourceType "StructureDefinition"
and derivation "constraint".

When fishing for FHIR in StructureDefinitionExporter, a not-yet-exported
Instance may be found. Export it using InstanceExporter before fishing
again.

* Find instances of extensions when fishing for extensions

An Instance in a FSHTank represents an Extension if it is a definition
instance of StructureDefinition, has a rule that assigns its derivation
as #constraint, and has a rule that assigns its type as Extension. An
Instance in a Package represents an Extensions if it is a definition
instance with resourceType "StructureDefinition", derivation
"constraint", and type "Extension".

Instances that represent Profiles must not have type "Extension".

* Find inline instances of extension when fishing for extensions

An inline instance that is an instance of Extension may be returned when
fishing in the FSHTank for extensions. Because they are inline
instances, they are not exported, so there is no change to fishing in
the Package.

* Find instances of code systems when fishing in package

A definitional instance of CodeSystem may be returned when fishing for
CodeSystem in a Package.

* Find instances of ValueSet when fishing for value sets

A definitional instance of ValueSet may be returned when fishing in the
FSHTank or Package for ValueSets.

* Do not return inline Instances of Extension when fishing for Extensions

An Instance of an Extension does not define an Extension, and therefore
is not a useful result when fishing for Extensions.

* Find instances when fishing for Logicals or Resources

An Instance of StructureDefinition may represent a Logical or a Resource
if it has the appropriate usage, derivation, and kind.

* Reorder find conditions

Check the name last, since other conditions can fail more quickly.