Check modifier extension usage on instances
When assigning a value on an instance, if the element receiving the
value is an extension or modifierExtension, try to check if the
extension is a modifier extension. If we can find the extension, show an
error when a modifier extension is used on an extension path, and show
an error when a non-modifier extension is used on a modifierExtension
path. If we can't find the extension, do not show an error.