Check modifier extension usage when constraining types
When an element is constrained with an OnlyRule, check to see if
modifier extensions are being used correctly. Log an error if a modifier
extension is being used on an extension path. The type constraints are
still applied.

Update modifier extension error messages for clarity and to include the
name of the relevant extension.