Check extensions used as path parts on Instances
When assigning values on Instances, it is possible to use an extension
as part of a path. When this is done, SUSHI automatically creates a
slice with that extension's url so that the rule's path will be valid.
When doing this, check if the extension is a modifier extension. Log an
error when a modifier extension is used on an extension element and when
a non-modifier extension is used on a modifierExtension element. Even
if these errors occur, the rule is still valid and its value is
assigned to the Instance.

Slices for extensions are automatically created on extension and
modifierExtension elements, since those are the two elements of type
Extension on standard FHIR resources.