Check if extension is or is not a modifier extension when constraining types or assigning instances (#960)
* Only modifier extensions should be used on modifierExtension elements

When handling a ContainsRule on an element whose only type is Extension,
check if the extension being added in the ContainsRule is a modifier
extension. A modifier extension is one where the root element has
isModifier set to true. Log an error when a modifier extension is used
on a path that ends in "extension" and when a non-modifier extension is
used on a path that ends in "modifierExtension".

* Check modifier extension usage on instances

When assigning a value on an instance, if the element receiving the
value is an extension or modifierExtension, try to check if the
extension is a modifier extension. If we can find the extension, show an
error when a modifier extension is used on an extension path, and show
an error when a non-modifier extension is used on a modifierExtension
path. If we can't find the extension, do not show an error.

* Check modifier extension usage when constraining types

When an element is constrained with an OnlyRule, check to see if
modifier extensions are being used correctly. Log an error if a modifier
extension is being used on an extension path. The type constraints are
still applied.

Update modifier extension error messages for clarity and to include the
name of the relevant extension.

* Tests for multiple types, constraining modifierExtension

* Check extensions used as path parts on Instances

When assigning values on Instances, it is possible to use an extension
as part of a path. When this is done, SUSHI automatically creates a
slice with that extension's url so that the rule's path will be valid.
When doing this, check if the extension is a modifier extension. Log an
error when a modifier extension is used on an extension element and when
a non-modifier extension is used on a modifierExtension element. Even
if these errors occur, the rule is still valid and its value is
assigned to the Instance.

Slices for extensions are automatically created on extension and
modifierExtension elements, since those are the two elements of type
Extension on standard FHIR resources.