Don't use HTTP HEAD to determine package availability (#975)
The packages.fhir.org server no longer allows HTTP HEAD requests.  Since SUSHI used HTTP HEAD to determine if a package was available on packages.fhir.org, this is obviously a problem.

Now we will just attempt the download on packages.fhir.org first and if there is an error, redo it on packages2.fhir.org.